changes for review comments.

+++ operators/tokenizer/bpe_decoder.hpp
@@ -125,7 +125,7 @@ struct KernelBpeDecoder : public BaseKernel {
       if (added_tokens_.count(token)) {
         const std::string ws = added_tokens_.at(token);
         decoded_token = (std::string)ws;
-      } else if (token < arr_vocab_.size()){
+      } else if (static_cast<size_t>(token) < arr_vocab_.size()){
         const auto str = arr_vocab_[token];
         for (auto wchr : str) {
           unsigned char uchr = byte_decoder_.at(wchr);
